Moved all PTransforms and interfaces into the SingleStoreIO class
