--- sdks/java/io/singlestore/src/main/java/org/apache/beam/sdk/io/singlestore/RowMapper.java
@@ -1,31 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.beam.sdk.io.singlestore;
-
-import java.io.Serializable;
-import java.sql.ResultSet;
-import org.apache.beam.sdk.values.PCollection;
-
-/**
- * An interface used by {@link Read} and {@link ReadWithPartitions} for converting each row of the
- * {@link ResultSet} into an element of the resulting {@link PCollection}.
- */
-@FunctionalInterface
-public interface RowMapper<T> extends Serializable {
-  T mapRow(ResultSet resultSet) throws Exception;
-}
