+++ sdks/java/io/singlestore/src/main/java/org/apache/beam/sdk/io/singlestore/SingleStoreIO.java
@@ -17,7 +17,48 @@
  */
 package org.apache.beam.sdk.io.singlestore;
 
+import static org.apache.beam.vendor.guava.v26_0_jre.com.google.common.base.Preconditions.checkArgument;
+import static org.apache.beam.vendor.guava.v26_0_jre.com.google.common.base.Preconditions.checkNotNull;
+
+import com.google.auto.value.AutoValue;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.PipedInputStream;
+import java.io.PipedOutputStream;
+import java.io.Serializable;
+import java.nio.charset.StandardCharsets;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.Statement;
+import java.util.ArrayList;
+import java.util.List;
+import javax.sql.DataSource;
+import org.apache.beam.sdk.coders.Coder;
+import org.apache.beam.sdk.coders.ListCoder;
+import org.apache.beam.sdk.coders.StringUtf8Coder;
+import org.apache.beam.sdk.io.range.OffsetRange;
+import org.apache.beam.sdk.transforms.Create;
+import org.apache.beam.sdk.transforms.DoFn;
+import org.apache.beam.sdk.transforms.Filter;
+import org.apache.beam.sdk.transforms.PTransform;
+import org.apache.beam.sdk.transforms.ParDo;
+import org.apache.beam.sdk.transforms.Reshuffle;
+import org.apache.beam.sdk.transforms.SerializableFunctions;
+import org.apache.beam.sdk.transforms.View;
+import org.apache.beam.sdk.transforms.display.DisplayData;
+import org.apache.beam.sdk.transforms.splittabledofn.RestrictionTracker;
+import org.apache.beam.sdk.transforms.windowing.GlobalWindow;
+import org.apache.beam.sdk.util.Preconditions;
+import org.apache.beam.sdk.values.PBegin;
 import org.apache.beam.sdk.values.PCollection;
+import org.apache.beam.sdk.values.PCollectionView;
+import org.apache.commons.dbcp2.BasicDataSource;
+import org.apache.commons.dbcp2.DelegatingStatement;
+import org.checkerframework.checker.nullness.qual.Nullable;
+import org.joda.time.Instant;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * IO to read and write data on SingleStoreDB.
@@ -124,7 +165,7 @@ public class SingleStoreIO {
    * @param <T> Type of the data to be read.
    */
   public static <T> Read<T> read() {
-    return new AutoValue_Read.Builder<T>().setOutputParallelization(true).build();
+    return new AutoValue_SingleStoreIO_Read.Builder<T>().setOutputParallelization(true).build();
   }
 
   /**
@@ -134,7 +175,7 @@ public class SingleStoreIO {
    * @param <T> Type of the data to be read.
    */
   public static <T> ReadWithPartitions<T> readWithPartitions() {
-    return new AutoValue_ReadWithPartitions.Builder<T>().build();
+    return new AutoValue_SingleStoreIO_ReadWithPartitions.Builder<T>().build();
   }
 
   /**
@@ -143,6 +184,772 @@ public class SingleStoreIO {
    * @param <T> Type of the data to be written.
    */
   public static <T> Write<T> write() {
-    return new AutoValue_Write.Builder<T>().build();
+    return new AutoValue_SingleStoreIO_Write.Builder<T>().build();
+  }
+
+  /**
+   * An interface used by {@link Read} and {@link ReadWithPartitions} for converting each row of the
+   * {@link ResultSet} into an element of the resulting {@link PCollection}.
+   */
+  @FunctionalInterface
+  public interface RowMapper<T> extends Serializable {
+    T mapRow(ResultSet resultSet) throws Exception;
+  }
+
+  /**
+   * An interface used by the SingleStoreIO {@link Read} to set the parameters of the {@link
+   * PreparedStatement}.
+   */
+  @FunctionalInterface
+  public interface StatementPreparator extends Serializable {
+    void setParameters(PreparedStatement preparedStatement) throws Exception;
+  }
+
+  /**
+   * An interface used by the SingleStoreIO {@link Write} to map a data from each element of {@link
+   * org.apache.beam.sdk.values.PCollection} to a List of Strings. Returned values are used to send
+   * LOAD DATA SQL queries to the database.
+   */
+  @FunctionalInterface
+  public interface UserDataMapper<T> extends Serializable {
+    List<String> mapRow(T element);
+  }
+
+  /**
+   * A POJO describing a SingleStoreDB {@link DataSource} by providing all properties needed to
+   * create it.
+   */
+  @AutoValue
+  public abstract static class DataSourceConfiguration implements Serializable {
+    abstract @Nullable String getEndpoint();
+
+    abstract @Nullable String getUsername();
+
+    abstract @Nullable String getPassword();
+
+    abstract @Nullable String getDatabase();
+
+    abstract @Nullable String getConnectionProperties();
+
+    abstract Builder builder();
+
+    @AutoValue.Builder
+    abstract static class Builder {
+      abstract Builder setEndpoint(String endpoint);
+
+      abstract Builder setUsername(String username);
+
+      abstract Builder setPassword(String password);
+
+      abstract Builder setDatabase(String database);
+
+      abstract Builder setConnectionProperties(String connectionProperties);
+
+      abstract DataSourceConfiguration build();
+    }
+
+    public static DataSourceConfiguration create(String endpoint) {
+      checkNotNull(endpoint, "endpoint can not be null");
+      return new AutoValue_SingleStoreIO_DataSourceConfiguration.Builder()
+          .setEndpoint(endpoint)
+          .build();
+    }
+
+    public DataSourceConfiguration withUsername(String username) {
+      checkNotNull(username, "username can not be null");
+      return builder().setUsername(username).build();
+    }
+
+    public DataSourceConfiguration withPassword(String password) {
+      checkNotNull(password, "password can not be null");
+      return builder().setPassword(password).build();
+    }
+
+    public DataSourceConfiguration withDatabase(String database) {
+      checkNotNull(database, "database can not be null");
+      return builder().setDatabase(database).build();
+    }
+
+    /**
+     * Sets the connection properties passed to driver.connect(...). Format of the string must be
+     * [propertyName=property;]*
+     *
+     * <p>NOTE - The "user" and "password" properties can be add via {@link #withUsername(String)},
+     * {@link #withPassword(String)}, so they do not need to be included here.
+     *
+     * <p>Full list of supported properties can be found here {@link <a
+     * href="https://docs.singlestore.com/managed-service/en/developer-resources/connect-with-application-development-tools/connect-with-java-jdbc/the-singlestore-jdbc-driver.html#connection-string-parameters">...</a>}
+     */
+    public DataSourceConfiguration withConnectionProperties(String connectionProperties) {
+      checkNotNull(connectionProperties, "connectionProperties can not be null");
+      return builder().setConnectionProperties(connectionProperties).build();
+    }
+
+    public static void populateDisplayData(
+        @Nullable DataSourceConfiguration dataSourceConfiguration, DisplayData.Builder builder) {
+      if (dataSourceConfiguration != null) {
+        builder.addIfNotNull(DisplayData.item("endpoint", dataSourceConfiguration.getEndpoint()));
+        builder.addIfNotNull(DisplayData.item("username", dataSourceConfiguration.getUsername()));
+        builder.addIfNotNull(DisplayData.item("database", dataSourceConfiguration.getDatabase()));
+        builder.addIfNotNull(
+            DisplayData.item(
+                "connectionProperties", dataSourceConfiguration.getConnectionProperties()));
+      }
+    }
+
+    public DataSource getDataSource() {
+      String endpoint = getEndpoint();
+      Preconditions.checkArgumentNotNull(endpoint, "endpoint can not be null");
+      String database = SingleStoreUtil.getArgumentWithDefault(getDatabase(), "");
+      String connectionProperties =
+          SingleStoreUtil.getArgumentWithDefault(getConnectionProperties(), "");
+      connectionProperties += (connectionProperties.isEmpty() ? "" : ";") + "allowLocalInfile=TRUE";
+      String username = getUsername();
+      String password = getPassword();
+
+      BasicDataSource basicDataSource = new BasicDataSource();
+      basicDataSource.setDriverClassName("com.singlestore.jdbc.Driver");
+      basicDataSource.setUrl(String.format("jdbc:singlestore://%s/%s", endpoint, database));
+
+      if (username != null) {
+        basicDataSource.setUsername(username);
+      }
+      if (password != null) {
+        basicDataSource.setPassword(password);
+      }
+      basicDataSource.setConnectionProperties(connectionProperties);
+
+      return basicDataSource;
+    }
+  }
+
+  /**
+   * A {@link PTransform} for reading data from SingleStoreDB. It is used by {@link
+   * SingleStoreIO#read()}.
+   */
+  @AutoValue
+  public abstract static class Read<T> extends PTransform<PBegin, PCollection<T>> {
+    private static final Logger LOG = LoggerFactory.getLogger(Read.class);
+
+    abstract @Nullable DataSourceConfiguration getDataSourceConfiguration();
+
+    abstract @Nullable String getQuery();
+
+    abstract @Nullable String getTable();
+
+    abstract @Nullable StatementPreparator getStatementPreparator();
+
+    abstract @Nullable Boolean getOutputParallelization();
+
+    abstract @Nullable RowMapper<T> getRowMapper();
+
+    abstract Builder<T> toBuilder();
+
+    @AutoValue.Builder
+    abstract static class Builder<T> {
+      abstract Builder<T> setDataSourceConfiguration(
+          DataSourceConfiguration dataSourceConfiguration);
+
+      abstract Builder<T> setQuery(String query);
+
+      abstract Builder<T> setTable(String table);
+
+      abstract Builder<T> setStatementPreparator(StatementPreparator statementPreparator);
+
+      abstract Builder<T> setOutputParallelization(Boolean outputParallelization);
+
+      abstract Builder<T> setRowMapper(RowMapper<T> rowMapper);
+
+      abstract Read<T> build();
+    }
+
+    public Read<T> withDataSourceConfiguration(DataSourceConfiguration config) {
+      checkNotNull(config, "dataSourceConfiguration can not be null");
+      return toBuilder().setDataSourceConfiguration(config).build();
+    }
+
+    public Read<T> withQuery(String query) {
+      checkNotNull(query, "query can not be null");
+      return toBuilder().setQuery(query).build();
+    }
+
+    public Read<T> withTable(String table) {
+      checkNotNull(table, "table can not be null");
+      return toBuilder().setTable(table).build();
+    }
+
+    public Read<T> withStatementPreparator(StatementPreparator statementPreparator) {
+      checkNotNull(statementPreparator, "statementPreparator can not be null");
+      return toBuilder().setStatementPreparator(statementPreparator).build();
+    }
+
+    /**
+     * Whether to reshuffle the resulting PCollection so results are distributed to all workers. The
+     * default is to parallelize and should only be changed if this is known to be unnecessary.
+     */
+    public Read<T> withOutputParallelization(Boolean outputParallelization) {
+      checkNotNull(outputParallelization, "outputParallelization can not be null");
+      return toBuilder().setOutputParallelization(outputParallelization).build();
+    }
+
+    public Read<T> withRowMapper(RowMapper<T> rowMapper) {
+      checkNotNull(rowMapper, "rowMapper can not be null");
+      return toBuilder().setRowMapper(rowMapper).build();
+    }
+
+    @Override
+    public PCollection<T> expand(PBegin input) {
+      DataSourceConfiguration dataSourceConfiguration = getDataSourceConfiguration();
+      Preconditions.checkArgumentNotNull(
+          dataSourceConfiguration, "withDataSourceConfiguration() is required");
+      RowMapper<T> rowMapper = getRowMapper();
+      Preconditions.checkArgumentNotNull(rowMapper, "withRowMapper() is required");
+      String actualQuery = SingleStoreUtil.getSelectQuery(getTable(), getQuery());
+
+      Coder<T> coder =
+          SingleStoreUtil.inferCoder(
+              rowMapper,
+              input.getPipeline().getCoderRegistry(),
+              input.getPipeline().getSchemaRegistry(),
+              LOG);
+
+      PCollection<T> output =
+          input
+              .apply(Create.of((Void) null))
+              .apply(
+                  ParDo.of(
+                      new ReadFn<>(
+                          dataSourceConfiguration,
+                          actualQuery,
+                          getStatementPreparator(),
+                          rowMapper)))
+              .setCoder(coder);
+
+      if (SingleStoreUtil.getArgumentWithDefault(getOutputParallelization(), true)) {
+        output = output.apply(new Reparallelize<>());
+      }
+
+      return output;
+    }
+
+    private static class ReadFn<ParameterT, OutputT> extends DoFn<ParameterT, OutputT> {
+      DataSourceConfiguration dataSourceConfiguration;
+      String query;
+      @Nullable StatementPreparator statementPreparator;
+      RowMapper<OutputT> rowMapper;
+
+      ReadFn(
+          DataSourceConfiguration dataSourceConfiguration,
+          String query,
+          @Nullable StatementPreparator statementPreparator,
+          RowMapper<OutputT> rowMapper) {
+        this.dataSourceConfiguration = dataSourceConfiguration;
+        this.query = query;
+        this.statementPreparator = statementPreparator;
+        this.rowMapper = rowMapper;
+      }
+
+      @ProcessElement
+      public void processElement(ProcessContext context) throws Exception {
+        DataSource dataSource = dataSourceConfiguration.getDataSource();
+        Connection conn = dataSource.getConnection();
+        try {
+          PreparedStatement stmt = conn.prepareStatement(query);
+          try {
+            if (statementPreparator != null) {
+              statementPreparator.setParameters(stmt);
+            }
+
+            ResultSet res = stmt.executeQuery();
+            try {
+              while (res.next()) {
+                context.output(rowMapper.mapRow(res));
+              }
+            } finally {
+              res.close();
+            }
+          } finally {
+            stmt.close();
+          }
+        } finally {
+          conn.close();
+        }
+      }
+    }
+
+    // Reparallelize PTransform is copied from JdbcIO
+    // https://github.com/apache/beam/blob/9d118bde5fe5a93c5f559ac3227758aea88185b8/sdks/java/io/jdbc/src/main/java/org/apache/beam/sdk/io/jdbc/JdbcIO.java#L2115
+    private static class Reparallelize<T> extends PTransform<PCollection<T>, PCollection<T>> {
+      @Override
+      public PCollection<T> expand(PCollection<T> input) {
+        // See https://issues.apache.org/jira/browse/BEAM-2803
+        // We use a combined approach to "break fusion" here:
+        // (see https://cloud.google.com/dataflow/service/dataflow-service-desc#preventing-fusion)
+        // 1) force the data to be materialized by passing it as a side input to an identity fn,
+        // then 2) reshuffle it with a random key. Initial materialization provides some parallelism
+        // and ensures that data to be shuffled can be generated in parallel, while reshuffling
+        // provides perfect parallelism.
+        // In most cases where a "fusion break" is needed, a simple reshuffle would be sufficient.
+        // The current approach is necessary only to support the particular case of SingleStoreIO
+        // where
+        // a single query may produce many gigabytes of query results.
+        PCollectionView<Iterable<T>> empty =
+            input
+                .apply("Consume", Filter.by(SerializableFunctions.constant(false)))
+                .apply(View.asIterable());
+        PCollection<T> materialized =
+            input.apply(
+                "Identity",
+                ParDo.of(
+                        new DoFn<T, T>() {
+                          @ProcessElement
+                          public void process(ProcessContext c) {
+                            c.output(c.element());
+                          }
+                        })
+                    .withSideInputs(empty));
+        return materialized.apply(Reshuffle.viaRandomKey());
+      }
+    }
+
+    @Override
+    public void populateDisplayData(DisplayData.Builder builder) {
+      super.populateDisplayData(builder);
+
+      DataSourceConfiguration.populateDisplayData(getDataSourceConfiguration(), builder);
+      builder.addIfNotNull(DisplayData.item("query", getQuery()));
+      builder.addIfNotNull(DisplayData.item("table", getTable()));
+      builder.addIfNotNull(
+          DisplayData.item(
+              "statementPreparator", SingleStoreUtil.getClassNameOrNull(getStatementPreparator())));
+      builder.addIfNotNull(DisplayData.item("outputParallelization", getOutputParallelization()));
+      builder.addIfNotNull(
+          DisplayData.item("rowMapper", SingleStoreUtil.getClassNameOrNull(getRowMapper())));
+    }
+  }
+
+  /**
+   * A {@link PTransform} for reading data from SingleStoreDB. It is used by {@link
+   * SingleStoreIO#readWithPartitions()}. {@link ReadWithPartitions} is preferred over {@link Read}
+   * because of the performance reasons.
+   */
+  @AutoValue
+  public abstract static class ReadWithPartitions<T> extends PTransform<PBegin, PCollection<T>> {
+    private static final Logger LOG = LoggerFactory.getLogger(ReadWithPartitions.class);
+
+    abstract @Nullable DataSourceConfiguration getDataSourceConfiguration();
+
+    abstract @Nullable String getQuery();
+
+    abstract @Nullable String getTable();
+
+    abstract @Nullable RowMapper<T> getRowMapper();
+
+    abstract @Nullable Integer getInitialNumReaders();
+
+    abstract Builder<T> toBuilder();
+
+    @AutoValue.Builder
+    abstract static class Builder<T> {
+      abstract Builder<T> setDataSourceConfiguration(
+          DataSourceConfiguration dataSourceConfiguration);
+
+      abstract Builder<T> setQuery(String query);
+
+      abstract Builder<T> setTable(String table);
+
+      abstract Builder<T> setRowMapper(RowMapper<T> rowMapper);
+
+      abstract Builder<T> setInitialNumReaders(Integer initialNumReaders);
+
+      abstract ReadWithPartitions<T> build();
+    }
+
+    public ReadWithPartitions<T> withDataSourceConfiguration(DataSourceConfiguration config) {
+      checkNotNull(config, "dataSourceConfiguration can not be null");
+      return toBuilder().setDataSourceConfiguration(config).build();
+    }
+
+    public ReadWithPartitions<T> withQuery(String query) {
+      checkNotNull(query, "query can not be null");
+      return toBuilder().setQuery(query).build();
+    }
+
+    public ReadWithPartitions<T> withTable(String table) {
+      checkNotNull(table, "table can not be null");
+      return toBuilder().setTable(table).build();
+    }
+
+    public ReadWithPartitions<T> withRowMapper(RowMapper<T> rowMapper) {
+      checkNotNull(rowMapper, "rowMapper can not be null");
+      return toBuilder().setRowMapper(rowMapper).build();
+    }
+
+    /** Pre-split initial restriction and start initialNumReaders reading at the very beginning. */
+    public ReadWithPartitions<T> withInitialNumReaders(Integer initialNumReaders) {
+      checkNotNull(initialNumReaders, "initialNumReaders can not be null");
+      return toBuilder().setInitialNumReaders(initialNumReaders).build();
+    }
+
+    @Override
+    public PCollection<T> expand(PBegin input) {
+      DataSourceConfiguration dataSourceConfiguration = getDataSourceConfiguration();
+      Preconditions.checkArgumentNotNull(
+          dataSourceConfiguration, "withDataSourceConfiguration() is required");
+      String database = dataSourceConfiguration.getDatabase();
+      Preconditions.checkArgumentNotNull(
+          database,
+          "withDatabase() is required for DataSourceConfiguration in order to perform readWithPartitions");
+      RowMapper<T> rowMapper = getRowMapper();
+      Preconditions.checkArgumentNotNull(rowMapper, "withRowMapper() is required");
+
+      int initialNumReaders = SingleStoreUtil.getArgumentWithDefault(getInitialNumReaders(), 1);
+      checkArgument(
+          initialNumReaders >= 1, "withInitialNumReaders() should be greater or equal to 1");
+
+      String actualQuery = SingleStoreUtil.getSelectQuery(getTable(), getQuery());
+
+      Coder<T> coder =
+          SingleStoreUtil.inferCoder(
+              rowMapper,
+              input.getPipeline().getCoderRegistry(),
+              input.getPipeline().getSchemaRegistry(),
+              LOG);
+
+      return input
+          .apply(Create.of((Void) null))
+          .apply(
+              ParDo.of(
+                  new ReadWithPartitions.ReadWithPartitionsFn<>(
+                      dataSourceConfiguration,
+                      actualQuery,
+                      database,
+                      rowMapper,
+                      initialNumReaders)))
+          .setCoder(coder);
+    }
+
+    private static class ReadWithPartitionsFn<ParameterT, OutputT>
+        extends DoFn<ParameterT, OutputT> {
+      DataSourceConfiguration dataSourceConfiguration;
+      String query;
+      String database;
+      RowMapper<OutputT> rowMapper;
+      int initialNumReaders;
+
+      ReadWithPartitionsFn(
+          DataSourceConfiguration dataSourceConfiguration,
+          String query,
+          String database,
+          RowMapper<OutputT> rowMapper,
+          int initialNumReaders) {
+        this.dataSourceConfiguration = dataSourceConfiguration;
+        this.query = query;
+        this.database = database;
+        this.rowMapper = rowMapper;
+        this.initialNumReaders = initialNumReaders;
+      }
+
+      @ProcessElement
+      public void processElement(
+          ProcessContext context, RestrictionTracker<OffsetRange, Long> tracker) throws Exception {
+        DataSource dataSource = dataSourceConfiguration.getDataSource();
+        Connection conn = dataSource.getConnection();
+        try {
+          for (long partition = tracker.currentRestriction().getFrom();
+              tracker.tryClaim(partition);
+              partition++) {
+            PreparedStatement stmt =
+                conn.prepareStatement(
+                    String.format("SELECT * FROM (%s) WHERE partition_id()=%d", query, partition));
+            try {
+              ResultSet res = stmt.executeQuery();
+              try {
+                while (res.next()) {
+                  context.output(rowMapper.mapRow(res));
+                }
+              } finally {
+                res.close();
+              }
+            } finally {
+              stmt.close();
+            }
+          }
+        } finally {
+          conn.close();
+        }
+      }
+
+      @GetInitialRestriction
+      public OffsetRange getInitialRange(@Element ParameterT element) throws Exception {
+        return new OffsetRange(0L, getNumPartitions());
+      }
+
+      @SplitRestriction
+      public void splitRange(
+          @Element ParameterT element,
+          @Restriction OffsetRange range,
+          OutputReceiver<OffsetRange> receiver) {
+        long numPartitions = range.getTo() - range.getFrom();
+        checkArgument(
+            initialNumReaders <= numPartitions,
+            "withInitialNumReaders() should not be greater then number of partitions in the database.\n"
+                + String.format(
+                    "InitialNumReaders is %d, number of partitions in the database is %d",
+                    initialNumReaders, range.getTo()));
+
+        for (int i = 0; i < initialNumReaders; i++) {
+          receiver.output(
+              new OffsetRange(
+                  range.getFrom() + numPartitions * i / initialNumReaders,
+                  range.getFrom() + numPartitions * (i + 1) / initialNumReaders));
+        }
+      }
+
+      private int getNumPartitions() throws Exception {
+        DataSource dataSource = dataSourceConfiguration.getDataSource();
+        Connection conn = dataSource.getConnection();
+        try {
+          Statement stmt = conn.createStatement();
+          try {
+            ResultSet res =
+                stmt.executeQuery(
+                    String.format(
+                        "SELECT num_partitions FROM information_schema.DISTRIBUTED_DATABASES WHERE database_name = %s",
+                        SingleStoreUtil.escapeString(database)));
+            try {
+              if (!res.next()) {
+                throw new Exception("Failed to get number of partitions in the database");
+              }
+
+              return res.getInt(1);
+            } finally {
+              res.close();
+            }
+          } finally {
+            stmt.close();
+          }
+        } finally {
+          conn.close();
+        }
+      }
+    }
+
+    @Override
+    public void populateDisplayData(DisplayData.Builder builder) {
+      super.populateDisplayData(builder);
+
+      DataSourceConfiguration.populateDisplayData(getDataSourceConfiguration(), builder);
+      builder.addIfNotNull(DisplayData.item("query", getQuery()));
+      builder.addIfNotNull(DisplayData.item("table", getTable()));
+      builder.addIfNotNull(
+          DisplayData.item("rowMapper", SingleStoreUtil.getClassNameOrNull(getRowMapper())));
+      builder.addIfNotNull(DisplayData.item("initialNumReaders", getInitialNumReaders()));
+    }
+  }
+
+  /**
+   * A {@link PTransform} for writing data to SingleStoreDB. It is used by {@link
+   * SingleStoreIO#write()}.
+   */
+  @AutoValue
+  public abstract static class Write<T> extends PTransform<PCollection<T>, PCollection<Integer>> {
+
+    private static final int DEFAULT_BATCH_SIZE = 100000;
+    private static final int BUFFER_SIZE = 524288;
+
+    abstract @Nullable DataSourceConfiguration getDataSourceConfiguration();
+
+    abstract @Nullable String getTable();
+
+    abstract @Nullable Integer getBatchSize();
+
+    abstract @Nullable UserDataMapper<T> getUserDataMapper();
+
+    abstract Builder<T> toBuilder();
+
+    @AutoValue.Builder
+    abstract static class Builder<T> {
+      abstract Builder<T> setDataSourceConfiguration(
+          DataSourceConfiguration dataSourceConfiguration);
+
+      abstract Builder<T> setTable(String table);
+
+      abstract Builder<T> setBatchSize(Integer batchSize);
+
+      abstract Builder<T> setUserDataMapper(UserDataMapper<T> userDataMapper);
+
+      abstract Write<T> build();
+    }
+
+    public Write<T> withDataSourceConfiguration(DataSourceConfiguration config) {
+      checkNotNull(config, "dataSourceConfiguration can not be null");
+      return toBuilder().setDataSourceConfiguration(config).build();
+    }
+
+    public Write<T> withTable(String table) {
+      checkNotNull(table, "table can not be null");
+      return toBuilder().setTable(table).build();
+    }
+
+    public Write<T> withUserDataMapper(UserDataMapper<T> userDataMapper) {
+      checkNotNull(userDataMapper, "userDataMapper can not be null");
+      return toBuilder().setUserDataMapper(userDataMapper).build();
+    }
+
+    /**
+     * Provide a maximum number of rows that is written by one SQL statement. Default is 100000.
+     *
+     * @param batchSize maximum number of rows that is written by one SQL statement
+     */
+    public Write<T> withBatchSize(Integer batchSize) {
+      checkNotNull(batchSize, "batchSize can not be null");
+      return toBuilder().setBatchSize(batchSize).build();
+    }
+
+    @Override
+    public PCollection<Integer> expand(PCollection<T> input) {
+      DataSourceConfiguration dataSourceConfiguration = getDataSourceConfiguration();
+      Preconditions.checkArgumentNotNull(
+          dataSourceConfiguration, "withDataSourceConfiguration() is required");
+      String table = getTable();
+      Preconditions.checkArgumentNotNull(table, "withTable() is required");
+      UserDataMapper<T> userDataMapper = getUserDataMapper();
+      Preconditions.checkArgumentNotNull(userDataMapper, "withUserDataMapper() is required");
+      int batchSize = SingleStoreUtil.getArgumentWithDefault(getBatchSize(), DEFAULT_BATCH_SIZE);
+      checkArgument(batchSize > 0, "batchSize should be greater then 0");
+
+      return input
+          .apply(
+              ParDo.of(
+                  new DoFn<T, List<String>>() {
+                    @ProcessElement
+                    public void processElement(ProcessContext context) throws Exception {
+                      context.output(userDataMapper.mapRow(context.element()));
+                    }
+                  }))
+          .setCoder(ListCoder.of(StringUtf8Coder.of()))
+          .apply(ParDo.of(new BatchFn<>(batchSize)))
+          .apply(ParDo.of(new WriteFn(dataSourceConfiguration, table)));
+    }
+
+    private static class BatchFn<ParameterT> extends DoFn<ParameterT, Iterable<ParameterT>> {
+      List<ParameterT> batch = new ArrayList<>();
+      int batchSize;
+
+      BatchFn(int batchSize) {
+        this.batchSize = batchSize;
+      }
+
+      @ProcessElement
+      public void processElement(ProcessContext context) {
+        batch.add(context.element());
+        if (batch.size() >= batchSize) {
+          context.output(batch);
+          batch = new ArrayList<>();
+        }
+      }
+
+      @FinishBundle
+      public void finish(FinishBundleContext context) {
+        if (batch.size() > 0) {
+          context.output(batch, Instant.now(), GlobalWindow.INSTANCE);
+          batch = new ArrayList<>();
+        }
+      }
+    }
+
+    private static class WriteFn extends DoFn<Iterable<List<String>>, Integer> {
+      DataSourceConfiguration dataSourceConfiguration;
+      String table;
+
+      WriteFn(DataSourceConfiguration dataSourceConfiguration, String table) {
+        this.dataSourceConfiguration = dataSourceConfiguration;
+        this.table = table;
+      }
+
+      @ProcessElement
+      public void processElement(ProcessContext context) throws Exception {
+        DataSource dataSource = dataSourceConfiguration.getDataSource();
+
+        Connection conn = dataSource.getConnection();
+        try {
+          Statement stmt = conn.createStatement();
+          try (PipedOutputStream baseStream = new PipedOutputStream();
+              InputStream inputStream = new PipedInputStream(baseStream, BUFFER_SIZE)) {
+            ((com.singlestore.jdbc.Statement) ((DelegatingStatement) stmt).getInnermostDelegate())
+                .setNextLocalInfileInputStream(inputStream);
+
+            final Exception[] writeException = new Exception[1];
+
+            Thread dataWritingThread =
+                new Thread(
+                    new Runnable() {
+                      @Override
+                      public void run() {
+                        try {
+                          Iterable<List<String>> rows = context.element();
+                          for (List<String> row : rows) {
+                            for (int i = 0; i < row.size(); i++) {
+                              String cell = row.get(i);
+                              if (cell.indexOf('\\') != -1) {
+                                cell = cell.replace("\\", "\\\\");
+                              }
+                              if (cell.indexOf('\n') != -1) {
+                                cell = cell.replace("\n", "\\n");
+                              }
+                              if (cell.indexOf('\t') != -1) {
+                                cell = cell.replace("\t", "\\t");
+                              }
+
+                              baseStream.write(cell.getBytes(StandardCharsets.UTF_8));
+                              if (i + 1 == row.size()) {
+                                baseStream.write('\n');
+                              } else {
+                                baseStream.write('\t');
+                              }
+                            }
+                          }
+
+                          baseStream.close();
+                        } catch (IOException e) {
+                          writeException[0] = e;
+                        }
+                      }
+                    });
+
+            dataWritingThread.start();
+            Integer rows =
+                stmt.executeUpdate(
+                    String.format(
+                        "LOAD DATA LOCAL INFILE '###.tsv' INTO TABLE %s",
+                        SingleStoreUtil.escapeIdentifier(table)));
+
+            context.output(rows);
+            dataWritingThread.join();
+
+            if (writeException[0] != null) {
+              throw writeException[0];
+            }
+          } finally {
+            stmt.close();
+          }
+        } finally {
+          conn.close();
+        }
+      }
+    }
+
+    @Override
+    public void populateDisplayData(DisplayData.Builder builder) {
+      super.populateDisplayData(builder);
+
+      DataSourceConfiguration.populateDisplayData(getDataSourceConfiguration(), builder);
+      builder.addIfNotNull(DisplayData.item("table", getTable()));
+      builder.addIfNotNull(DisplayData.item("batchSize", getBatchSize()));
+      builder.addIfNotNull(
+          DisplayData.item(
+              "userDataMapper", SingleStoreUtil.getClassNameOrNull(getUserDataMapper())));
+    }
   }
 }
