--- sdks/java/io/singlestore/src/main/java/org/apache/beam/sdk/io/singlestore/UserDataMapper.java
@@ -1,31 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.beam.sdk.io.singlestore;
-
-import java.io.Serializable;
-import java.util.List;
-
-/**
- * An interface used by the SingleStoreIO {@link Write} to map a data from each element of {@link
- * org.apache.beam.sdk.values.PCollection} to a List of Strings. Returned values are used to send
- * LOAD DATA SQL queries to the database.
- */
-@FunctionalInterface
-public interface UserDataMapper<T> extends Serializable {
-  List<String> mapRow(T element);
-}
